backend/python/requirements.txt
pixivpy3==3.7.3
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
